fix(sem): error with `var`/`lent`-returning templates (#1018)
## Summary

Fix using templates and macros with a `var T` or `lent T` return type
leading to compiler crashes or wrong code being generated. Both the C
and the VM backend were affected.

## Details

When the receiving type is a `var` or `lent` type, `sigmatch` produced
implicit conversions to said type. These are wrong at the type-level,
as a conversion cannot turn a non-view value into a view. Due a
separate issue with `semAfterMacroCall`, this led to semantically
invalid expressions being produced after expansion of the
aforementioned templates.

The `var` and `lent` modifiers are now dropped from the produced
conversions, both fixing the concrete problem with templates and
allowing for the removal of several workarounds in the compiler. 

Reporting a warning for narrowing object conversions where information
is lost relied on the `var` modifier in conversion. The test is
replaced with looking for the problematic conversion in receiver
contexts (except for `discard`, as the value isn't used there).

Finally, a general test is added for ensuring that templates returning
`var` or `lent` types behave the same as a normal procedure would. They
currently don't, but most of the test at least compiles and runs now.