fix(sem): template/macro issues with `var`/`lent` (#1006)
## Summary

* fix the result of `lent` returning macros and templates being
  assignable when the final body of the template/macro ends in a
  mutable lvalue expression
* fix wrong code being generated for templates and macros that had a
  concrete tuple return type and where the body was of `var`/`lent`
  tuple type (all backends were affected)

In terms of how the result can be used, macros and templates with
`var`/`lent` return types now behave the same as the other routines do.

Fixes https://github.com/nim-works/nimskull/issues/328.

## Details

The problem was that the expansion of a template/macro using a concrete
type as the return type wasn't typed correctly:
* `semExpr` doesn't insert a hidden deref for `var`/`lent`-typed
  expressions, leading to a `var`/`lent`-typed expression being passed
  to `fitNode`
* `semExpr` doesn't enforce that the expression is not `void`-typed
  (i.e., a statement)
* `fitNode` alone doesn't handle `var` and `lent` target types
  correctly

This led to semantics inconsistent with `var`/`lent`-returning
procedures and code generation problems.

For fixing the issues:
* `semExpr` is replaced with `semExprWithType`
* `fitNode` is replaced with `fitNodeConsiderViewType`

Two templates in the `results` module can now use concrete types as the
return type again.