mir: improve the `Changeset` API
Summary
=======

Remove the cursor tracking and require the modification position to be
manually provided when recording a change. This makes the API less
stateful and also significantly reduces the time and memory cost of
creating and `Changeset`.

Details
=======

**Concrete changes:**
* the `seek` and `position` routines are removed
* the `pos` and `tree` field are removed
* all routines that need to perform a tree traversal now have a
  `MirTree` parameter
* all routines for recording changes now have a `NodePosition`
  parameter

The reason for the high creation overhead was that the `Changeset`
carried a full copy of the MIR tree it recorded the changes for. This
was necessary for providing automatic cursor movement, which was
intended to efficiently record changes to nodes following immediately
after each other.

However, there was only a single use of the automatic cursor movement
functionality, and it's unlikely that this is going to change. The
feature is therefore removed.