make the `in` magic work the same across backends (#949)
## Summary

Make range-checks of the left-hand operand in built-in `in` calls
affect the operations result instead of causing a `RangeDefect`: if the
operand is not in the value range expected by the `set`, the call
returns 'false', regardless of the selected backend and whether a
literal set construction is used or not.

The `ContainsConv` warning that warned about the previous behaviour is
removed.

## Details

* lower `x in y` where `x` requires a range check (`nkChckRange`) into
  `a <= x and x <= y and x in y` in `mirgen`
* the lowering has to happen in `mirgen`, as non-local control-flow
  needs to be fixed at MIR generation and the otherwise emitted
  `chckRange` call has control-flow effects
* remove the `mChckRange` handling from `ccgexprs.genInOp`
* remove the `ContainsConv` warning and the associated test
* add a test for the magic's new behaviour regarding range checks

---------

Co-authored-by: Clyybber <darkmine956@gmail.com>