mirexec: integrate data-flow instructions
Instead of the hybrid traversal, where the control-flow part was
handled through the CFG and the data-flow part through the MIR, both
are now combined into the CFG, making it a data-flow graph.

Apart from slightly simplifying the traversal logic, this is also more
efficient and, most importantly, the MIR doesn't have to support the
pseudo-execution anymore. This is going to allow for significant
changes to the MIR's structure.

The module's documentation still needs to be adjusted.