rework the alias analysis
Update the alias analysis to work with the new MIR. For more efficient
comparisons, paths now need to be compiled into a more fitting
representation first, which also allows for following aliases.

This is much more powerful than the previous analysis, since it, in
terms of the legacy MIR, allows for following region parameters.