mirgen: update to the new IR structure
Change to `mirgen` to emit the revised MIR code. Large changes are
required to the code gen for expressions, as the structure is vastly
different.

Lots of fixing and iterating already took place, but there's still a
lot left to do, most notably cleaning up, adjusting the documentation,
and restoring proper source mappings.

`transf` is changed to no longer omit the to-openArray conversions;
doing so only complicated `mirgen`. In addition, `transf` for now
patches to-varargs conversions into to-openArray conversions. `mirgen`
would have to do so otherwise, but it is ill-equipped to do so at the
moment (no access to a ``IdGenerator``).

Finally, some small additions to the MIR were required:
* `mnkDefUnpack` was restored, it's still needed until destructors are
  injected by `mirgen` for now
* for easier transformation by later passes, aliases and the bind
  statements are introduced. Using `var T` and `lent T` locals wouldn't
  work, as creating those requires access to an `IdGenerator`, which
  neither `mirgen` nor the transformation passes have