mirexec: update to the new IR structure
Due to the new structure, all instruction are now associated with a
*statement*. This renders a lot of the assumptions made by the data-
flow analysis stale, meaning that much of it has to fundamentally
change.

More decision making is moved into the data-flow graph creation.
Most notably, `def` are now also emitted for definition statements,
which is going to simplify some of the data-flow analysis a bit.