mirexec: use instruction spans
Specify the active span and start position to the traversal iterators
in instruction indices rather than in node indices. This simplifies the
iterators, is more efficient, and it's also a requirement for the
planned instruction-centric `injectdestructors` processing.

The `rangeFor` and `instructions` routine are provided for translating
a node span into an instruction span and for iterating over all data-
flow instructions.