fix(cgen): wrong code being generated for `finish`
This is a pre-existing issue. A statement like:
```nim
  var v = finished(x)
```

had no code generated for the assignment. This was caused by the
`genBreakState` only setting the rope of the destination loc (partially
correct for empty locs, wrong for filled locs).

The procedure is fixed by using `putIntoDest` (which properly handles
non-empty locs). In addition, the procedure is moved to `ccgexprs`, as
it performs code generation for a magic call.