restore `echo` support
All code generators expected an array construction expression as the
operand, but this isn't the case anymore (the array is now assigned to
a temporary first).

To resolve this, the string operands are now passed to the echo magic
as arguments directly. The C backend needs an array, so the type of the
would-be array is passed along for `cgen` to use when manually emitting
the array setup code.