update the `injectdestructors` pass
Much of transformation part is rewritten in order to accommodate the
new IR. Some parts simplify while other become a bit more complex, but
there's an overall slowdown. The new IR isn't tailored to the
`injectdestructors` pass as much as the old one was, so this is
expected.

The `genMarkCyclic` call injection is removed for now, as it
introduces additional overhead for no benefit. It's an unfinished
optimization that can only be enabled by modifying the ORC sources,
meaning that no code relied on or used it.

Finally, the `Changeset` API is extended with the `replaceSingle`
procedure.