mirconstr: update to the new IR
The DSL is completely obsolete now and thus removed. Instead,
`mirconstr` now provides the `MirBuilder` type together with routines
that are very similar to the ones used in `mirgen`.

Ideally, `mirgen` would use `MirBuilder` too, but that would require
supporting the front vs. staging buffer distinction in `MirBuilder`,
which is not so easy to support efficiently.

For reasons of ergonomics, `MirBuilder` is able to allocate
temporaries, but this currently requires an integration into
`Changeset`, which is a layering violation.