fix run-time crash with `var openArray`
The temporary receiving a mutable `openArray` created via conversion
didn't reach `cgen` typed as `var openArray`, which caused
`prepareStrMutation` to not be injected (and thus segmentation faults
at run-time).

`mirgen` now makes sure that the `ToSlice` produced for a
`(HiddenAddr (HiddenStdConv x))` tree is typed as `var openArray`. This
also allows for only treating the creation of mutable views/slices
as mutations in the data-flow graph.