restore copies of internal object fields
For copying the type fields of objects, `liftdestructors` previously
produced the following assignment:
```nim
mAccessTypeField(x) = mAccessTypeField(y)
```
effectively treating the magic call as a template expanded by
the code generator.

The new MIR doesn't support magic being used like that anymore, and so
a different solution is required: the assignment is generalized into
the `mCopyInternal` magic, which `cgen` then expands into the type field
copy.

A similar problem to the one detailed above also exists for the
`mAccessEnv` magic.