mirexec: integrate data-flow instructions
Instead of the hybrid traversal, where the control-flow part was
handled through the CFG and the data-flow part through the MIR, both
are now combined into the CFG, making it a data-flow graph.

Apart from simplifying the traversal logic, this is also more efficient
and, most importantly, the MIR doesn't have to support the pseudo-execution
anymore.

Another simplification this allows for is removing the special casing
around assignment-on-definitions in `analysis`/`injectdestructors`, as
those assignments are now simply represented as a `def` in the data-
flow graph.