mirconstr: working `SourceId` attachments
Make the `apply` mechanism internal only and provide `setSource` in its
place. The `setSource` is simpler, more intuitive, and also fixes the
remaining association issues (e.g., moving a staged fragment to the
final buffer where the final buffer wasn't fully processed yet) that
weren't solvable with the old approach.

In addition, changeset recording is no longer responsible for updating
the `info` fields when recording multi-node changes, allowing callsites
to use arbitrary source IDs.

The downside is that AST nodes have to be eagerly added to the source
map now, resulting in obsolete mappings if no `MirNode` actually uses
their ID. To reduce the amount of pressure on the `SourceMap`,
`useSource` only registers an AST node if its different from the active
one, which prevents a large number of redundant mappings.

In the end, the number of total nodes registered with `SourceMap`s
increases by around 13% (measured during bootstrapping).