unify the MIR tree creation
Merge the double-buffering functionality used in `mirgen` into
`MirBuilder`. This:
* allows for using `MirBuilder` in `mirgen`
* removes the logic duplication between `mirgen` and `mirconstr`
* allows for better incremental AST -> MIR translation, as the
  `generateCode` overload now takes a `MirBuilder` instance

In terms of code movement / architectural changes:
* `MirBuffer` and its associated routines move to `mirconstr`
* `commit` is replaced with a slightly higher-level `push`/`pop`
  interface

The double buffering is now more hidden, with a structured interface
for switching between the staging/final buffer. This does complicate
`mirgen` a bit, but it's necessary in order to not force using the
staging buffer in contexts where it's not needed (e.g., lowering
transformation).

Source attachments still don't quite work yet, but it's getting there.

Misc:
* a `mirgen` bug with `skField` symbols used with the `.emit` pragma
  is fixed
* dynlib-loader production works properly again
* initializers for lifted global work again