mirgen: proper expression kind detection
For now, the better detection allows for eliminating an unnecessary
temporary that was created when assigning a call expression to a named
location in a `var`/`let` statement (e.g., `let len = len(x)`).

In the future (read, separate PR), the expression kind detection is
going to be used for deciding where to use a `move`, `copy`, or `sink`.