mir: allow lvalue expressions as operands
It's now possible to use a full lvalue expression as an operand to a
call or construction, making the MIR more intuitive and much more
flexible. For example, the future optimization pass for forwarding
eliminating temporaries where safe can now be implemented as a MIR ->
MIR pass.

`mirgen` is likely only going to benefit little from this change, as
it still needs to go through temporaries in most cases.