mirgen: better by-name argument translation
With lvalue expression now supported in call operand positions, less
aliases are required, reducing the amount of MIR nodes. If an
expression is stable, it can be used directly as the argument
expression.

In addition, a translation issue with the first argument to
`lent`-returning procedures is fixed: calls are no longer bound to
aliases. This would happen if the first argument to a `lent`-returning
call is a call rvalue.