vm: remove `openArray` conversion support
Summary
=======

Implement the conversion in the code generator instead, making the VM
slightly simpler. A VM crash when assigning an `openArray` parameter to
an `openArray` field is also fixed.

Details
=======

* implement the current to-openArray conversion (essentially a
  `seq` copy) in `vmgen` by using the existing opcodes
* remove the built-in VM `openArray` conversion support (in
  `opConv`)

Having the operation implemented in bytecode makes the VM leaner and
reduces the amount of run-time dispatching.

In addition, the operation is now also performed when the source
expression is already an `openArray`. Locals and parameters of
`openArray` type may also be stored as arrays or strings, in which
case directly assigning the value via `WrLoc` would result in a run-
time type error.