MIR: refactor the data-flow analysis (#1031)
## Summary

Change the data-flow analysis implementation used by the MIR phase
(`mirexec`) to rely less on the MIR's tree structure, preparing for
changes to the latter. This also speeds up the traversal and simplifies
the implementation.

## Details

The MIR data-flow analysis used a hybrid approach so far: `mirexec`
provided control-flow graph traversal routines for traversing the basic
blocks in a MIR tree, with the data-flow relevant effects directly
computed from the MIR nodes during traversal.

### Problems

This approach worked well, but it has the problem of necessitating that
MIR nodes appear in the order their data-flow effects happen, which
is what ultimately led to the current instruction-like design of the
MIR. An exception was already made for calls and assignments (refer to
`computeValuesAndEffects`), where mutation effects have to be applied
when the *call/assignment* node is visited, not when the argument's
*tag* node (which comes earlier) is visited.

### Revised implementation

In order to make evolving the MIR's layout possible without having to
change the data-flow analysis algorithms, the data-flow effects are now
encoded directly in the control-flow graph, making it a data-flow
graph.

**Data-flow operations**:
* `opDef`: corresponds to `ekReassign` effects and is also used for
  definitions (`mnkDef`)
* `opUse`: represents all lvalue usages that are not consumes
* `opConsume`: corresponds to `mnkConsume` and other consume contexts
* `opKill`, `opMutate`, `opInvalidate`: correspond to the
  `EffectKind` of the same name
* `opMutateGlobal`: represents a mutation of an unspecified global

The data-flow instruction are emitted together with the control-flow
instructions in `computeCfg`.

**Changes:**
* move the `ConsumeCtx`, `UseContext`, `OpsWithEffects` sets into
  `mirtrees`
* instead of iterating over the tree and having to map the positions to
  control-flow instructions, the `traverse`, `traverseReverse`, and
  `traverseFromExits` iterators only need to iterate the instruction
  list (which renders the `tree` parameters obsolete)
* the iterators return the data-flow operation kind plus expression
  directly

**Changes to the usage sites:**
* change the `traverse` iterator to start at `start` rather than at
  `start + 1`. This is more intuitive and easier to implement. The
  transitive callsites `isLastRead`, `isLastWrite`, `computeOwnership`,
  and `needsReset` are adjusted accordingly, by applying the offset
  themselves
* support `start` pointing outside the provided `span` in `traverse`,
  making the API easier to use (e.g., no extra guard logic is needed)
* remove the special handling for `mnkDef` statements from
  `analysis.computeAlive` -- the information is now encoded in the
  data-flow graph
* remove the now-obsolete effect tracking for `Call`, `Asgn`, etc.
  nodes
* the `tmir_exec.nim` is updated to work with the `mirexec` changes

`ControlFlowGraph` is renamed to `DataFlowGraph`, `computeCfg` to
`computeDfg`, and the associated documentation is updated to reflect
the shift in purpose.

### Additional benefits

* computing the data-flow operations in a single place (i.e., during
  data-flow graph create) means that downstream users don't have to
  duplicate the logic
* the traversal routines no longer need a `MirTree`, making it much
  easier to generalize them into something that can be used without the
  MIR