vm: remove `openArray` conversion support (#1033)
## Summary

Implement the conversion operation in bytecode instead, making the VM
slightly simpler. A VM crash when assigning an `openArray` parameter to
an `openArray` field is also fixed.

## Details

* implement the current to-openArray conversion (essentially a
  `seq` copy) in `vmgen` as `SetLenSeq` + `ArrCopy`, replicating
  the VM's implementation in `opConv`
* remove the built-in VM `openArray` conversion support from
  `opConv` 

Having the operation implemented in bytecode makes the VM leaner and
reduces the amount of run-time dispatching.

In addition, the operation is now also performed when the source
expression is already an `openArray`. Locals and parameters of
`openArray` type may also be stored as arrays or strings, in which
case directly assigning the value via `WrLoc` would result in a run-
time type error.