fix(cgen): run-time crash when modifying a `seq`
Summary
=======

Fix a crash when trying to add to, set the length of, or assign to a
sequence part of a tuple or array location that was initialized with a
fully constant tuple or array construction expression.

Details
=======

Constructions of empty sequences were treated as literals during the
MIR phase, which allowed eliding the `=copy` hook call (for backends
that need it) for them.

In `cgen`, constructions of empty `seq`s are then expanded into
a payload allocation setup code (`ccgexprs.genSeqConstr`). However,
fully constant tuple or array construction expression are lifted
into C constants, which includes empty `seq` constructions part of the
expression.

If the `seq` in the variable the C constant was assigned to is then
modified (through an `add`), a write into the immutable data section
takes place, crashing the program.

`mirgen` is changed to no longer special empty `seq` construction,
ensuring that a `=copy` is always used, thus subsequently preventing
tuple/array constructions they're part of to be lifted into C
constants. So that no now-unnecessary `newSeqPayload` call is emitted
by `cgen`, empty `seq` constructions are now also lifted into C
constants.

Alternatively, `seqs_v2.add`, `seqs_v2.grow`, and `seqs_v2.setLen`
could be patched to properly take the `strlitFlag` into account, which
would allow for constant sequences to still be lifted into C constants.
It would also introduce a small amount of overhead to each of the
aforementioned operations, and is thus decided against.