internal: improve the `mirexec` API (#1034)
## Summary

Refactor the `traverse` iterators to accept instructions spans/indices
and provide translation routines for node -> instruction translation.
This slightly simplifies the implementation, allows for instruction-
level addressing, and it reduces the amount of node-position ->
instruction translation.

## Details

It's not necessary for the `traverse` iterator to perform the node
position -> instruction translation, and since multiple instructions
may be associated with a single node, this also limited the
granularity at which traversal could start/end to nodes.

The `Subgraph` type is added, representing a sub-graph (i.e.,
instruction span) within the data-flow graph. Computing the `Subgraph`
and translating a node to an instruction position are done via
`subgraphFor` and `find`. All three `traverse` iterators are changed
to accept both a `Subgraph` and `InstrPos`. 

The change in API is propagated through the routines in the `analysis`
module, allowing for performing the node span -> `Subgraph` translation
only once for each analyzed location (instead of at the start of each
traversal). Instead of the enclosing `mnkScope`'s sub-tree,
`EntityInfo` stores the pre-computed `Subgraph` -- `getAliveRange` is
renamed to `getAliveGraph` and returns the subgraph directly.

Finally, the `traverse` iterators are changed to return a
`DataFlowOpcode` rather than an `Opcode` value, eliminating unnecessary
`case` branches when matching over the value.