js: improve cast-to-`pointer` support (#1040)
## Summary

Support casting `ref` of aggregates, `proc`, and scalar types to
`pointer` values and the pointer back to the original type, for the JS
backend. Using the result of these casts previously either worked,
misbehaved at run-time, or caused an internal compiler error.

## Details

The `jsgen.genCast` procedure didn't account for all non-fat-pointer
types in `genCast`. If the `TCompRes` resulting from such a cast
reached `genVarInit`, the assertion for the `etyBaseIndex` handling
triggered.

As an interim solution, `genCast` now handles all to- and from-
`pointer` casts the same:
* non-`etyBaseIndex` values are cast into `pointer` by treating the
  expression as the address and `null` as the index (already happened
  for `etyObject`)
* `pointer` values are cast into non-`etyBaseIndex` values by using the
  address part as the expression (the inverse of the to-`pointer` cast)
* `pointer` to fat-pointer casts and vice versa are no-ops

This implementation means that casting `A` to `pointer` and the
`pointer` to `B` is undefined behaviour for the JS target, but that was
already the case before the change.

### Tests

Test coverage for casts is increased by:
* adding a test for casting `ref`-of-scalars to `pointer` and back
* adding a test for casting `ref`-of-aggregate to `pointer` and back
* adding a test for casting an `int` to `pointer` and back

The not-yet-working `misc/tcast.nim` is marked as a known issue for the
JS target.