mirgen: move `genPath` back into `genx`
This reduces code duplication and fixes:
* path nodes not using the correct origin information
* procedure literals used as the operand of `cast`, `conv`, etc.
  crashing the compiler

`genPath` is replaced with `genOperand`. The big difference between
`genOperand` and `genPath` is that `genOperand` bases its decision on
what to do on the MIR expression rather than on the AST, making it more
robust.

In addition, the target of a dereference is now always captured when
non-pure, fixing an evaluation order violation for `x[]` used as a call
argument.