injectdestructors: fix `Switch` lowering
A binding from a `mnkPathVariant` expression was created, which is
effectively undefined behaviour at the moment.

To make the produced MIR code well-behaving, the alias is created from
a `mnkPathNamed` pointing to the discriminator. The `mnkPathVariant`
expression still needs to be passed to the `=destroy` call, otherwise
the path invalidation wouldn't work.