fix: ORC crash with closures
A bug introduced with the MIR rework. Closure environment references
collected through `=trace` would cause the cycle collector to crash.

The cause of the crash is with `liftdestructors` producing a
`nimTraceRef(addr accessEnv(cl))` statement for tracing the closure,
thus effectively relying on the `mAccessEnv` to be treated as a code-
generator-expanded template.

Since calls (including magic calls) may only appear as rvalues, the
closure's reference was first assigned to a temporary of which the
address was then taken, resulting in an illegal stack access.

`mirgen` now lowers the `mAccessEnv` magic into a tuple access,
resolving the issue, but also requiring `cgen` to detect this special
access and translate it separately. Once closures are lowered into
tuples during the MIR phase, the special handling in `cgen` can
be removed again.