regression: disable parts of `tassignment.nim`
The issue is a known and pre-existing one, but it's surfacing in this
specific test now that all non-pure argument expressions are assigned
to temporaries first.

`x` being passed directly to `raiseEx` disabled RVO for the call, but
since is no longer the case, RVO is enabled for the call, thus causing
the mutation on the result variable to be observable to the outside.