regression: disable `tsplit_into_table.nim`
Assigning to an `openArray` lvalue that has side-effect (i.e., can
raise) requires a binding the lvalue to an alias, but the C code
generator doesn't support this at the moment.

This is going to be resolved once the side-effects of lvalues can be
captured independently from the lvalue (i.e., by handling index checks
at the MIR level).