cgen: don't emit unnecessary type definitions
Summary
=======

Don't emit the full C type definition for a type when assigning a `ref`
or `ptr` of said type, reducing the amount of generated C code and thus
the amount of work for both the NimSkull and C compiler.

Details
=======

The full C type definition of the pointed-to type was requested on
assignment (`ccgstmts.genAsgn`) of the pointer-like type, but this is
unnecessary, as C doesn't require the pointed-to type to be a complete
type when assigning the pointer.

It's likely that this is a leftover from when `genDeref` didn't request
the type definition of the pointed-to type. Requesting the type
definition is removed, as translating the left and right side of the
assignment already pulls in the relevant type definitions.