analysis: store ownership status in a `PackedSet`
Whether an lvalue is movable from is a boolean value now, meaning that
the table is unnecessary -- a `PackedSet` is used instead. The now-
obsolete `Owned` enum is also removed.

As a consequence, `solveOwnership` can now return an instance of
`Values` rather than having to mutate one.

Finally, the un-used `Value` parameters is removed from both
`isLastRead` and `isLastWrite`.